BubbleSortExample image.
