+++ TemplateMethod/src/main/java/devslp/designpatterns/TemplateMethod/BubbleSortExample/BubbleSorter.java
@@ -1,4 +1,4 @@
-package devslp.designpatterns.TemplateMethod.SortMethodTemplateExample;
+package devslp.designpatterns.TemplateMethod.BubbleSortExample;
 
 public abstract class BubbleSorter {
     protected int length;
