+++ TemplateMethod/src/test/java/devslp/designpatterns/TemplateMethod/BubbleSortExample/BubbleSorterTest.java
@@ -1,4 +1,4 @@
-package devslp.designpatterns.TemplateMethod.SortMethodTemplateExample;
+package devslp.designpatterns.TemplateMethod.BubbleSortExample;
 
 
 import org.junit.Test;
